+++ buildpacks/dotnet/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2025-05-09
+
 ### Changed
 
 - Enhance error message for missing project references in solution files to include common causes and detailed resolution steps. ([#259](https://github.com/heroku/buildpacks-dotnet/pull/259))
@@ -185,7 +187,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation.
 
-[unreleased]: https://github.com/heroku/buildpacks-dotnet/compare/v0.5.0...HEAD
+[unreleased]: https://github.com/heroku/buildpacks-dotnet/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/heroku/buildpacks-dotnet/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/heroku/buildpacks-dotnet/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/heroku/buildpacks-dotnet/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/heroku/buildpacks-dotnet/compare/v0.3.6...v0.4.0